Add support for views which emit null as a key.
According to <http://wiki.apache.org/couchdb/HTTP_view_API#Creating_Views>,
it's legal for a view to emit null as a key, presumably if it doesn't care
about the output order, and the "key" value in the output is a real null
in this case, not the string "null".

Browser-Couch tried to use the emitted key value as a dictionary key,
which converts null to the string "null", resulting in incorrect output,
and potentially ambiguous if the map function happened to emit some rows
with the key "null" as well.

Now it supports real null keys and outputs them as such.